translation_service/requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
openai
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.23.0
requests